+++ src/extension.ts
@@ -1,6 +1,7 @@
-// The module 'vscode' contains the VS Code extensibility API
+// The module 'vscode' contains the VSCode extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import { VSCodeMarketplaceClient } from 'vscode-marketplace-client';
 
 // Import the Configs, Controllers, and Providers
 import {
@@ -66,12 +67,13 @@ export async function activate(context: vscode.ExtensionContext) {
       config.update(workspaceConfig);
 
       if (isEnabled) {
-        const message = vscode.l10n.t('{0} is now enabled and ready to use', [
-          EXTENSION_DISPLAY_NAME,
-        ]);
+        const message = vscode.l10n.t(
+          'The {0} extension is now enabled and ready to use',
+          [EXTENSION_DISPLAY_NAME],
+        );
         vscode.window.showInformationMessage(message);
       } else {
-        const message = vscode.l10n.t('{0} is now disabled', [
+        const message = vscode.l10n.t('The {0} extension is now disabled', [
           EXTENSION_DISPLAY_NAME,
         ]);
         vscode.window.showInformationMessage(message);
@@ -111,12 +113,12 @@ export async function activate(context: vscode.ExtensionContext) {
         title: vscode.l10n.t('Release Notes'),
       },
       {
-        title: vscode.l10n.t('Close'),
+        title: vscode.l10n.t('Dismiss'),
       },
     ];
 
     const message = vscode.l10n.t(
-      'New version of {0} is available. Check out the release notes for version {1}',
+      "The {0} extension has been updated. Check out what's new in version {1}",
       [EXTENSION_DISPLAY_NAME, currentVersion],
     );
     vscode.window.showInformationMessage(message, ...actions).then((option) => {
@@ -144,6 +146,58 @@ export async function activate(context: vscode.ExtensionContext) {
   }
 
   // -----------------------------------------------------------------
+  // Check for updates
+  // -----------------------------------------------------------------
+
+  // Check for updates to the extension
+  try {
+    // Retrieve the latest version
+    VSCodeMarketplaceClient.getLatestVersion(
+      USER_PUBLISHER,
+      EXTENSION_NAME,
+    ).then((latestVersion) => {
+      // Check if the latest version is different from the current version
+      if (latestVersion !== currentVersion) {
+        const actions: vscode.MessageItem[] = [
+          {
+            title: vscode.l10n.t('Update Now'),
+          },
+          {
+            title: vscode.l10n.t('Dismiss'),
+          },
+        ];
+
+        const message = vscode.l10n.t(
+          'A new version of {0} is available. Update to version {1} now',
+          [EXTENSION_DISPLAY_NAME, latestVersion],
+        );
+        vscode.window
+          .showInformationMessage(message, ...actions)
+          .then(async (option) => {
+            if (!option) {
+              return;
+            }
+
+            // Handle the actions
+            switch (option?.title) {
+              case actions[0].title:
+                await vscode.commands.executeCommand(
+                  'workbench.extensions.action.install.anotherVersion',
+                  `${USER_PUBLISHER}.${EXTENSION_NAME}`,
+                );
+                break;
+
+              default:
+                break;
+            }
+          });
+      }
+    });
+  } catch (error) {
+    console.error('Error retrieving extension version:', error);
+  }
+
+  // -----------------------------------------------------------------
   // Register FilesController
   // -----------------------------------------------------------------
 
@@ -156,7 +210,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -173,7 +227,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -190,7 +244,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);